requirements.txt
aniso8601==10.0.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-RESTful==0.3.10
fsspec==2024.12.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
protobuf==5.29.3
psycopg2-binary==2.9.10
pyotp==2.9.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
safetensors==0.5.2
scipy==1.15.1
sentencepiece==0.2.0
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.37
sympy==1.13.1
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
tweepy==4.15.0
twikit==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
